requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
idna==3.4
oauthlib==3.2.1
phonenumbers==8.12.54
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
urllib3==1.26.12
